+++ src/lib.rs
@@ -62,7 +62,7 @@ assert!(htmlize::unescape("3 &times 4 &gt; 10") == "3 × 4 > 10");
 //!   * `unescape_fast`: provide fast version of [`unescape()`]. This does _not_
 //!     enable the `entities` feature automatically.
 //!
-//!     This takes perhaps 30 seconds longer to build than `unescape`, but the
+//!     This takes perhaps 2 seconds longer to build than `unescape`, but the
 //!     performance is significantly better in the worst cases. That said, the
 //!     performance of of the `unescape` version is already pretty good, so I
 //!     don’t recommend enabling this unless you really need it.