+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## main branch
 
 * Clarify examples in documentation and README.
+* Fix a few spelling mistakes in documentation.
 
 ## Release 1.0.6 (2025-04-26)